requirements.txt
python-telegram-bot~=22.0
config~=0.5.1
SQLAlchemy~=2.0.40
psycopg2-binary==2.9.7
python-dotenv==1.0.0